Suppress warnings & formatting changes (#2473)
* Make external library headers as system-headers so that the compiler does
  not print the warnings during compilation
* Fix various warnings inside ArrayFire
* Fix formatting

(cherry picked from commit bdef3da0af442a4c34ce4f06e1c7150bb569090f)

+++ src/backend/cpu/kernel/random_engine.hpp
@@ -37,65 +37,65 @@ static const double PI_VAL =
 #define DBL_FACTOR ((1.0) / (UINTLMAX + (1.0)))
 #define HALF_DBL_FACTOR ((0.5) * DBL_FACTOR)
 
-template <typename T>
+template<typename T>
 T transform(uint *val, int index) {
     T *oval = (T *)val;
     return oval[index];
 }
 
-template <>
+template<>
 char transform<char>(uint *val, int index) {
     char v = val[index >> 2] >> (8 << (index & 3));
     v      = (v & 0x1) ? 1 : 0;
     return v;
 }
 
-template <>
+template<>
 uchar transform<uchar>(uint *val, int index) {
     uchar v = val[index >> 2] >> (index << 3);
     return v;
 }
 
-template <>
+template<>
 ushort transform<ushort>(uint *val, int index) {
     ushort v = val[index >> 1] >> (16 << (index & 1));
     return v;
 }
 
-template <>
+template<>
 short transform<short>(uint *val, int index) {
     return transform<ushort>(val, index);
 }
 
-template <>
+template<>
 uint transform<uint>(uint *val, int index) {
     return val[index];
 }
 
-template <>
+template<>
 int transform<int>(uint *val, int index) {
     return transform<uint>(val, index);
 }
 
-template <>
+template<>
 uintl transform<uintl>(uint *val, int index) {
     uintl v = (((uintl)val[index << 1]) << 32) | ((uintl)val[(index << 1) + 1]);
     return v;
 }
 
-template <>
+template<>
 intl transform<intl>(uint *val, int index) {
     return transform<uintl>(val, index);
 }
 
 // Generates rationals in [0, 1)
-template <>
+template<>
 float transform<float>(uint *val, int index) {
     return 1.f - (val[index] * FLT_FACTOR + HALF_FLT_FACTOR);
 }
 
 // Generates rationals in [0, 1)
-template <>
+template<>
 double transform<double>(uint *val, int index) {
     uintl v = transform<uintl>(val, index);
     return 1.0 - (v * DBL_FACTOR + HALF_DBL_FACTOR);
@@ -112,27 +112,27 @@ double transform<double>(uint *val, int index) {
 // ELEMS_PER_ITER correspond to elementsPerBlock in the CUDA backend, so each
 // "iter" (iteration) here correspond to a CUDA thread block doing its work.
 // This change was prompted by issue #2429
-template <typename T>
+template<typename T>
 void philoxUniform(T *out, size_t elements, const uintl seed, uintl counter) {
     uint hi  = seed >> 32;
     uint lo  = seed;
     uint hic = counter >> 32;
     uint loc = counter;
 
-    constexpr int RESET_CTR = MAX_RESET_CTR_VAL / sizeof(T);
-    constexpr int ELEMS_PER_ITER =
+    constexpr size_t RESET_CTR = MAX_RESET_CTR_VAL / sizeof(T);
+    constexpr size_t ELEMS_PER_ITER =
         WRITE_STRIDE * 4 * sizeof(uint) / sizeof(T);
 
     int num_iters = divup(elements, ELEMS_PER_ITER);
-    int len       = num_iters * ELEMS_PER_ITER;
+    size_t len    = num_iters * ELEMS_PER_ITER;
 
-    constexpr int NUM_WRITES = 16 / sizeof(T);
-    for (int iter = 0; iter < len; iter += ELEMS_PER_ITER) {
-        for (int i = 0; i < WRITE_STRIDE; i += RESET_CTR) {
-            for (int j = 0; j < RESET_CTR; ++j) {
+    constexpr size_t NUM_WRITES = 16 / sizeof(T);
+    for (size_t iter = 0; iter < len; iter += ELEMS_PER_ITER) {
+        for (size_t i = 0; i < WRITE_STRIDE; i += RESET_CTR) {
+            for (size_t j = 0; j < RESET_CTR; ++j) {
                 // first_write_idx is the first of the 4 locations that will
                 // be written to
-                ptrdiff_t first_write_idx = iter + i + j;
+                uintptr_t first_write_idx = iter + i + j;
                 if (first_write_idx >= elements) { break; }
 
                 // Recalculate key and ctr to emulate how the CUDA backend
@@ -144,8 +144,8 @@ void philoxUniform(T *out, size_t elements, const uintl seed, uintl counter) {
 
                 // Use the same ctr array for each of the 4 locations,
                 // but each of the location gets a different ctr value
-                for (int buf_idx = 0; buf_idx < NUM_WRITES; ++buf_idx) {
-                    int out_idx = iter + buf_idx * WRITE_STRIDE + i + j;
+                for (size_t buf_idx = 0; buf_idx < NUM_WRITES; ++buf_idx) {
+                    size_t out_idx = iter + buf_idx * WRITE_STRIDE + i + j;
                     if (out_idx < elements) {
                         out[out_idx] = transform<T>(ctr, buf_idx);
                     }
@@ -158,7 +158,7 @@ void philoxUniform(T *out, size_t elements, const uintl seed, uintl counter) {
 #undef MAX_RESET_CTR_VAL
 #undef WRITE_STRIDE
 
-template <typename T>
+template<typename T>
 void threefryUniform(T *out, size_t elements, const uintl seed, uintl counter) {
     uint hi     = seed >> 32;
     uint lo     = seed;
@@ -178,7 +178,7 @@ void threefryUniform(T *out, size_t elements, const uintl seed, uintl counter) {
     }
 }
 
-template <typename T>
+template<typename T>
 void boxMullerTransform(T *const out1, T *const out2, const T r1, const T r2) {
     /*
      * The log of a real value x where 0 < x < 1 is negative.
@@ -201,7 +201,7 @@ void boxMullerTransform(uint val[4], float *temp) {
                        transform<float>(val, 3));
 }
 
-template <typename T>
+template<typename T>
 void philoxNormal(T *out, size_t elements, const uintl seed, uintl counter) {
     uint hi     = seed >> 32;
     uint lo     = seed;
@@ -220,7 +220,7 @@ void philoxNormal(T *out, size_t elements, const uintl seed, uintl counter) {
     }
 }
 
-template <typename T>
+template<typename T>
 void threefryNormal(T *out, size_t elements, const uintl seed, uintl counter) {
     uint hi     = seed >> 32;
     uint lo     = seed;
@@ -245,7 +245,7 @@ void threefryNormal(T *out, size_t elements, const uintl seed, uintl counter) {
     }
 }
 
-template <typename T>
+template<typename T>
 void uniformDistributionMT(T *out, size_t elements, uint *const state,
                            const uint *const pos, const uint *const sh1,
                            const uint *const sh2, uint mask,
@@ -270,7 +270,7 @@ void uniformDistributionMT(T *out, size_t elements, uint *const state,
     state_write(state, l_state);
 }
 
-template <typename T>
+template<typename T>
 void normalDistributionMT(T *out, size_t elements, uint *const state,
                           const uint *const pos, const uint *const sh1,
                           const uint *const sh2, uint mask,
@@ -297,7 +297,7 @@ void normalDistributionMT(T *out, size_t elements, uint *const state,
     state_write(state, l_state);
 }
 
-template <typename T>
+template<typename T>
 void uniformDistributionCBRNG(T *out, size_t elements,
                               af_random_engine_type type, const uintl seed,
                               uintl counter) {
@@ -313,7 +313,7 @@ void uniformDistributionCBRNG(T *out, size_t elements,
     }
 }
 
-template <typename T>
+template<typename T>
 void normalDistributionCBRNG(T *out, size_t elements,
                              af_random_engine_type type, const uintl seed,
                              uintl counter) {
